+++ src/main/java/rlti/com/rh/funcionario/application/api/request/CargoRequest.java
@@ -1,6 +1,6 @@
 package rlti.com.rh.funcionario.application.api.request;
 
-import rlti.com.rh.funcionario.domain.GrauDeInstrucao;
+import rlti.com.rh.funcionario.domain.enums.GrauDeInstrucao;
 
 public record CargoRequest(
         String nomeCargo,
