+++ src/main/java/rlti/com/rh/funcionario/domain/Formacao.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import rlti.com.rh.funcionario.domain.enums.SituacaoCurso;
 
 import java.time.LocalDate;
