+++ src/main/java/rlti/com/rh/funcionario/domain/SalarioBase.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import rlti.com.rh.funcionario.application.api.request.SalarioBaseRequest;
 
 import java.math.BigDecimal;
@@ -15,6 +16,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Getter
 @Setter
+@EntityListeners(AuditingEntityListener.class)
 @Entity(name = "SALARIO_BASE")
 public class SalarioBase {
     @Id
