+++ src/main/java/rlti/com/rh/funcionario/domain/enums/SituacaoCurso.java
@@ -1,4 +1,4 @@
-package rlti.com.rh.funcionario.domain;
+package rlti.com.rh.funcionario.domain.enums;
 
 public enum SituacaoCurso {
     CONCLUIDO,
